move 'identical' tests to separate file
