+++ types/ramda/ramda-tests.ts
@@ -1513,16 +1513,6 @@ type Pair = KeyValuePair<string, number>;
 };
 
 () => {
-    const o = {};
-    R.identical(o, o); // => true
-    R.identical(1, 1); // => true
-    R.identical("2", "1"); // => false
-    R.identical([], []); // => false
-    R.identical(0, -0); // => false
-    R.identical(NaN, NaN); // => true
-};
-
-() => {
     const testPath = ["x", 0, "y"];
     const testObj  = {x: [{y: 2, z: 3}, {y: 4, z: 5}]};
