+++ types/ramda/tsconfig.json
@@ -92,6 +92,7 @@
         "test/has-tests.ts",
         "test/hasIn-tests.ts",
         "test/hasPath-tests.ts",
-        "test/head-tests.ts"
+        "test/head-tests.ts",
+        "test/identical-tests.ts"
     ]
 }
